+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## v3.4.0-dev.0 (May 23, 2025)
 - Now compatible with either React 17 and 18 [#2506](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2506)
 
+- Refactor commerce-sdk-react to allow injecting ApiClients [#2519](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2519)
+
 ## v3.3.0 (May 22, 2025)
 
 - Fix inconsistency between dwsid and access token for guest login when hybrid authentication is enabled [#2397](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2397)